reducing quantity

+++ app/Services/CartService.php
@@ -48,8 +48,12 @@ class CartService extends TransformerService{
 			$cart = $this->getCart("session", $request->input('sessionId'));
 		}
 
-		$response = $this->cartProductService->removeFromCart($request, $cart);
-		
+		if ($request->has('quantity')) {
+			$response = $this->cartProductService->reduceQuantity($request, $cart);		
+		}else{
+			$response = $this->cartProductService->removeFromCart($request, $cart);
+		}
+
 		return $response;
 	}
 
@@ -154,7 +158,7 @@ class CartService extends TransformerService{
 				
 				if ($duplicatedCartProduct != null) {
 					// Increase quantity of original entry
-					$this->cartProductService->increaseQuantity($newCartProduct, $duplicatedCartProduct);
+					$this->cartProductService->mergeQuantity($newCartProduct, $duplicatedCartProduct);
 
 					// Delete duplicated cart product entry
 					$this->cartProductService->delete($duplicatedCartProduct);
